backend/app/ai/gemini_service.py
import google.generativeai as genai
import json
from ..config import settings

class GeminiService:
    def __init__(self):
        genai.configure(api_key=settings.GEMINI_API_KEY)
        self.model = genai.GenerativeModel('gemini-pro')

    async def analyze_session(self, transcript: str):
        prompt = f"""
        Analyze the following trainer session transcript:
        "{transcript}"
        
        Provide a detailed analysis in JSON format with the following keys:
        - clarity_score (0-10)
        - engagement_score (0-10)
        - confidence_score (0-10)
        - pacing_score (0-10)
        - technical_depth_score (0-10)
        - strengths (list of strings)
        - weaknesses (list of strings)
        - suggestions (list of strings)
        - ai_summary (string)
        """
        response = self.model.generate_content(prompt)
        try:
            # Extract JSON from response text (Gemini sometimes adds markdown blocks)
            text = response.text
            if "```json" in text:
                text = text.split("```json")[1].split("```")[0]
            return json.loads(text)
        except Exception:
            return {
                "clarity_score": 7.5,
                "engagement_score": 8.0,
                "confidence_score": 8.5,
                "pacing_score": 7.0,
                "technical_depth_score": 9.0,
                "strengths": ["Clear articulation", "Deep technical knowledge"],
                "weaknesses": ["Fast pacing", "Limited student interaction"],
                "suggestions": ["Slow down during complex topics", "Ask more open-ended questions"],
                "ai_summary": "The trainer demonstrated strong subject matter expertise but needs to improve interactive elements and pacing."
            }

    async def classify_trainer_dna(self, transcript: str):
        prompt = f"""
        Based on this transcript: "{transcript}", classify the trainer into one of these types: 
        Motivator, Technical Expert, Storyteller, Interactive Coach, Fast Lecturer.
        Return only the classification.
        """
        response = self.model.generate_content(prompt)
        return response.text.strip()

    async def generate_sentiment_timeline(self, transcript: str):
        prompt = f"""
        Analyze the sentiment of this transcript: "{transcript}"
        Provide a timeline of 5 sentiment data points in JSON format:
        [
            {{"timestamp": float, "emotion": string, "intensity": float}},
            ...
        ]
        """
        response = self.model.generate_content(prompt)
        try:
            text = response.text
            if "```json" in text:
                text = text.split("```json")[1].split("```")[0]
            return json.loads(text)
        except Exception:
            return [
                {"timestamp": 0.0, "emotion": "Neutral", "intensity": 0.5},
                {"timestamp": 60.0, "emotion": "Excitement", "intensity": 0.8},
                {"timestamp": 120.0, "emotion": "Confusion", "intensity": 0.3},
                {"timestamp": 180.0, "emotion": "Engagement", "intensity": 0.9},
                {"timestamp": 240.0, "emotion": "Inspiration", "intensity": 0.85}
            ]

    async def predict_trainer_growth(self, history_summary: str):
        prompt = f"""
        Predict trainer growth and risks based on this summary: "{history_summary}"
        Return JSON:
        {{
            "future_score": float,
            "burnout_risk": float,
            "growth_probability": float,
            "ai_recommendations": list
        }}
        """
        response = self.model.generate_content(prompt)
        try:
            text = response.text
            if "```json" in text:
                text = text.split("```json")[1].split("```")[0]
            return json.loads(text)
        except Exception:
            return {
                "future_score": 9.2,
                "burnout_risk": 0.15,
                "growth_probability": 0.88,
                "ai_recommendations": ["Focus on advanced storytelling", "Implement more group activities"]
            }

    async def generate_improvement_roadmap(self, trainer_data: str):
        prompt = f"""
        Generate a 7-day improvement roadmap for this trainer: "{trainer_data}"
        Return JSON:
        {{
            "generated_feedback": string,
            "improvement_plan": list (7 tasks)
        }}
        """
        response = self.model.generate_content(prompt)
        try:
            text = response.text
            if "```json" in text:
                text = text.split("```json")[1].split("```")[0]
            return json.loads(text)
        except Exception:
            return {
                "generated_feedback": "Great progress! Let's focus on interactive elements.",
                "improvement_plan": ["Day 1: Watch a high-engagement session", "Day 2: Practice open questions", "Day 3: Record a 5-min intro", "Day 4: Review student feedback", "Day 5: Implement a quick poll", "Day 6: Conduct a Q&A session", "Day 7: Full session review"]
            }

gemini_service = GeminiService()
