backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
google-generativeai>=0.8.0
websockets>=13.0
aiofiles>=24.0
bcrypt>=4.2.0
